refactored get user posts

+++ src/main/java/gag/sasu/dto/GetAllPostsResponse.java
@@ -13,6 +13,6 @@ public class GetAllPostsResponse {
 
     List<VoteForPostResponse> posts;
 
-    int maxPosts;
+    long maxPosts;
 
 }
